RAGPIPELINE/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.1
langchain-astradb>=0.0.5
langchain-community>=0.0.1
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0
pypdf>=3.15.1
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.9
